rtika blog  (#166)
* Create 2018-03-27-rtika-introduction.md

creating blank file.

* Rename 2018-03-27-rtika-introduction.md to 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

slightly reducing the amount of code in the blog post

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

reduced size of headings

* Update 2018-04-03-rtika-introduction.md

updated date in yaml and made final edits. It is ready for final review.

* Update 2018-04-03-rtika-introduction.md

* Update 2018-04-03-rtika-introduction.md

merged in edits from @stefaniebutland, and changed date to today.

* Create 2018-04-25-rtika-introduction.md

updated blog post date embedded in file name

* Delete 2018-04-03-rtika-introduction.md

removing this file with outdated file name, to be replaced with updated file name

* Update 2018-04-25-rtika-introduction.md

fixed one link and added link to package at the very end of the article.